requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
colorama==0.4.6
Django==4.1.3
django-crispy-forms==1.14.0
exceptiongroup==1.0.4
factory-boy==3.2.1
Faker==15.3.4
h11==0.14.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
outcome==1.2.0
packaging==22.0
pluggy==1.0.0
psycopg2==2.9.5
pycparser==2.21
PySocks==1.7.1
pytest==7.2.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.13
wsproto==1.2.0
